# docs/openapi/ms-roadmap.yaml
openapi: 3.0.3
info:
  title: roadmap-service — Tema 10 · Grupo 10
  description: >
    Contrato de "Roadmap y Progreso". Fuente de verdad del front (ver
    path/01-arquitectura-y-stack.md §4 — contract-first). Cada path indica su estado:
    `[IMPLEMENTADO]` ya existe en el código, `[PLANEADO]` es el contrato objetivo de
    fases futuras — el front puede mockear contra él desde ya.
  version: "0.1.0"
servers:
  - url: /api/roadmap
    description: >
      Prefijo externo servido por el Gateway (deck Tema 01). El Gateway reenvía el path
      completo sin reescribir — este servidor es también el base path del controller.
...
paths:
...
  /roadmaps:
    post:
      summary: "[IMPLEMENTADO] Crear el roadmap de un curso"
      tags: [grafo]
      security: [{ bearerAuth: [PROFESOR] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cursoCohorteId]
              properties:
                cursoCohorteId: { type: string, format: uuid }
      responses:
        "201":
          description: Roadmap creado en estado BORRADOR
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Roadmap" }
        "409":
          description: Ya existe un roadmap activo para ese curso
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/ProblemDetail" }
  /roadmaps/{cc}:
    parameters:
      - $ref: "#/components/parameters/cursoCohorteId"
    get:
      summary: "[IMPLEMENTADO] Traer el grafo (editor) o filtrado por progreso (alumno)"
      tags: [grafo]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Roadmap" }
        "404":
          description: No existe roadmap para ese curso
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/ProblemDetail" }
    put:
      summary: "[PLANEADO] Editar metadata, cambiar borrador ↔ publicado"
      tags: [grafo]
      security: [{ bearerAuth: [PROFESOR] }]
      responses:
        "200": { description: OK }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Validado por el Gateway (401 si falla). Nosotros leemos X-User-Roles ya
        verificado y decidimos el negocio (403) — ver 06-contrato-api.md §0.3.
  parameters:
    cursoCohorteId:
      name: cc
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    ProblemDetail:
      type: object
      description: RFC 9457 — ver 06-contrato-api.md §0.5
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        traceId: { type: string }
    Roadmap:
      type: object
      properties:
        id: { type: string, format: uuid }
        cursoCohorteId: { type: string, format: uuid }
        estado: { type: string, enum: [BORRADOR, PUBLICADO] }
        creadoEn: { type: string, format: date-time }
        actualizadoEn: { type: string, format: date-time }
...